+++ theme-orange-country/src/main/java/com/orange/ouds/theme/orangecountry/OrangeCountryTheme.kt
@@ -13,7 +13,7 @@
 package com.orange.ouds.theme.orangecountry
 
 import com.orange.ouds.theme.orange.OrangeTheme
-import com.orange.ouds.theme.orangecountry.tokens.semantic.orangeCountrySemanticColors
+import com.orange.ouds.theme.orangecountry.tokens.orangeCountrySemanticColorTokens
 import com.orange.ouds.theme.tokens.semantic.OudsColorTokens
 
 class OrangeCountryTheme : OrangeTheme() {
@@ -21,6 +21,6 @@ class OrangeCountryTheme : OrangeTheme() {
     override val name: String
         get() = "Orange country"
 
-    override val colors: OudsColorTokens
-        get() = orangeCountrySemanticColors
+    override val colorTokens: OudsColorTokens
+        get() = orangeCountrySemanticColorTokens
 }
