+++ theme-orange/src/main/java/com/orange/ouds/theme/orange/OrangeTheme.kt
@@ -13,7 +13,7 @@
 package com.orange.ouds.theme.orange
 
 import com.orange.ouds.theme.OudsCustomTheme
-import com.orange.ouds.theme.orange.tokens.semantic.orangeSemanticColors
+import com.orange.ouds.theme.orange.tokens.orangeColorSemanticTokens
 import com.orange.ouds.theme.tokens.semantic.OudsColorTokens
 import kotlinx.parcelize.Parcelize
 
@@ -25,6 +25,6 @@ open class OrangeTheme : OudsCustomTheme {
     override val name: String
         get() = ORANGE_THEME_NAME
 
-    override val colors: OudsColorTokens
-        get() = orangeSemanticColors
+    override val colorTokens: OudsColorTokens
+        get() = orangeColorSemanticTokens
 }
