+++ theme-white-label/src/main/java/com/orange/ouds/theme/whitelabel/WhiteLabelTheme.kt
@@ -14,7 +14,7 @@ package com.orange.ouds.theme.whitelabel
 
 import com.orange.ouds.theme.OudsCustomTheme
 import com.orange.ouds.theme.tokens.semantic.OudsColorTokens
-import com.orange.ouds.theme.whitelabel.tokens.semantic.WhiteLabelSemanticColors
+import com.orange.ouds.theme.whitelabel.tokens.WhiteLabelColorSemanticTokens
 import kotlinx.parcelize.Parcelize
 
 @Parcelize
@@ -23,6 +23,6 @@ open class WhiteLabelTheme : OudsCustomTheme {
     override val name: String
         get() = "White label"
 
-    override val colors: OudsColorTokens
-        get() = WhiteLabelSemanticColors
+    override val colorTokens: OudsColorTokens
+        get() = WhiteLabelColorSemanticTokens
 }
